fix: stabilize native PF-001 quality gates

+++ apps/launcher/internal/adapters/process/executable_linux.go
@@ -33,7 +33,7 @@ func platformExecutableIdentity(info os.FileInfo) (string, error) {
 		return "", os.ErrInvalid
 	}
 	return executableIdentityText(
-		uint64(status.Dev), status.Ino, status.Ctim.Sec, status.Ctim.Nsec, info.Size(),
+		status.Dev, status.Ino, status.Ctim.Sec, status.Ctim.Nsec, info.Size(),
 	), nil
 }
